+++ src/typed_settings/processors.py
@@ -10,6 +10,18 @@ from .dict_utils import iter_settings, set_path
 from .types import OptionList, SettingsClass, SettingsDict
 
 
+__all__ = [
+    "Protocol",
+    "UrlHandler",
+    "UrlProcessor",
+    "handle_raw",
+    "handle_script",
+    "handle_op",
+    "FormatProcessor",
+    "JinjaProcessor",
+]
+
+
 LOGGER = logging.getLogger("typed_settings")