+++ src/typed_settings/types.py
@@ -4,12 +4,13 @@ Internal data structures.
 from collections.abc import Collection
 from enum import Enum
 from pathlib import Path
+from types import MappingProxyType
 from typing import (
     Any,
     Dict,
     Final,
     Generic,
-    List,
+    NamedTuple,
     Tuple,
     Type,
     TypeVar,
@@ -17,6 +18,32 @@ from typing import (
 )
 
 import attrs
+
+from ._compat import PY_39
+
+
+__all__ = [
+    "AUTO",
+    "SECRET_REPR",
+    "T",
+    "ET",
+    "ST",
+    "SettingsClass",
+    "SettingsInstance",
+    "SettingsDict",
+    "OptionName",
+    "OptionPath",
+    "OptionInfo",
+    "OptionList",
+    "OptionDict",
+    "LoaderMeta",
+    "LoadedValue",
+    "LoadedSettings",
+    "MergedSettings",
+    "Secret",
+    "SecretStr",
+    "SECRETS_TYPES",
+]
 
 
 SECRET_REPR: Final[str] = "*******"
@@ -27,25 +54,9 @@ ET = TypeVar("ET", bound=Enum)  # Enum type
 ST = TypeVar("ST", bound=attrs.AttrsInstance)  # SettingsInstance
 SettingsClass = Type[attrs.AttrsInstance]
 SettingsInstance = attrs.AttrsInstance
-SettingsDict = Dict[str, Union[Any, "SettingsDict"]]
-
-
-def _type2name(value: Union[str, Any]) -> str:
-    if isinstance(value, str):
-        return value
-    return type(value).__name__
-
-
-@attrs.frozen
-class LoaderMeta:
-    name: str = attrs.field(converter=_type2name)
-    cwd: Path = attrs.field(factory=Path.cwd)
-
-
-@attrs.frozen
-class LoadedSettings:
-    settings: SettingsDict
-    meta: LoaderMeta
+OptionName = str
+OptionPath = str
+SettingsDict = Dict[OptionName, Union[Any, "SettingsDict"]]
 
 
 class _Auto:
@@ -72,6 +83,15 @@ Sentinel to indicate the lack of a value when ``None`` is ambiguous.
 """
 
 
+def _type2name(value: Union[str, Any]) -> str:
+    """
+    Return either *value* if it is a str or else its type name.
+    """
+    if isinstance(value, str):
+        return value
+    return type(value).__name__
+
+
 @attrs.frozen
 class OptionInfo:
     """
@@ -80,7 +100,7 @@ class OptionInfo:
     Each instance represents a single attribute of an apps's settings class.
     """
 
-    path: str
+    path: OptionPath
     """
     Dotted path to the option name relative to the root settings class.
     """
@@ -92,20 +112,92 @@ class OptionInfo:
 
     cls: type
     """
-    The option's settings class.  This is either the root settings class or a
-    nested one.
+    The option's settings class.  This is either the root settings class or a nested
+    one.
     """
 
 
-OptionList = List[OptionInfo]
+OptionList = Tuple[OptionInfo, ...]
 """
 A flat list of all available options, including those from nested settings.
 """
 
-MergedSettings = Dict[str, Tuple[OptionInfo, LoaderMeta, Any]]
+if PY_39:
+    OptionDict = MappingProxyType[OptionPath, OptionInfo]
+else:
+    from typing import Mapping
+
+    OptionDict = Mapping[OptionPath, OptionInfo]  # type: ignore
 """
-A list of loaded settings values with the meta data of the loader that loaded the
-value.
+A dict version of :class:`OptionList`.
+"""
+
+
+@attrs.frozen
+class LoaderMeta:
+    """
+    Meta data about the loader that loaded a set of option values.
+
+    That data is used to improve error reporting and for converting loaded values to the
+    target type.
+    """
+
+    name: str = attrs.field(converter=_type2name)
+    """
+    The loader's name.  Can be a string or the loader class itself (it's class name is
+    will then be used).
+    """
+
+    cwd: Path = attrs.field(factory=Path.cwd)
+    """
+    The current working directory of the loader.
+
+    This can, e.g., be the actual cwd or the parent directory of a config file.  It
+    allows to resolve relative paths in option values based on the corresponding
+    loader's cwd.
+    """
+
+
+class LoadedValue(NamedTuple):
+    """
+    A container for a loaded option value and the meta data of the originating loader.
+    """
+
+    value: Any
+    """
+    The loaded option value.
+    """
+
+    loader_meta: LoaderMeta
+    """
+    Meta data of the loader that loaded the corresponding value.
+    """
+
+
+@attrs.frozen
+class LoadedSettings:
+    """
+    A container for the settings loaded by a single loader, and the meta data of that
+    loader.
+    """
+
+    settings: SettingsDict
+    """
+    The loaded settings values.
+    """
+
+    meta: LoaderMeta
+    """
+    Meta data of the loader that loaded the settings.
+    """
+
+
+MergedSettings = Dict[OptionPath, LoadedValue]
+"""
+A dict that maps a dotted option path to a loaded option value.
+
+The values may come from different loaders, so each option values stores the meta data
+of it's loader.
 """
 
 
@@ -113,8 +205,8 @@ class SecretStr(str):
     """
     A subclass of :class:`str` that masks the output of :func:`repr()`.
 
-    It is less secure than :class:`Secret` but is a drop-in replacement for
-    normal strings.
+    It is less secure than :class:`Secret` but is a drop-in replacement for normal
+    strings.
 
     The main use case is avoiding accidental secrets leakage via tracebacks.
     It also helps to enforce secret usage via Typing.
@@ -137,18 +229,17 @@ class Secret(Generic[T]):
     """
     A secret wrapper around any value.
 
-    It makes it very hard to accidentally leak the secret, even when printing
-    or logging it.
+    It makes it very hard to accidentally leak the secret, even when printing or logging
+    it.
 
-    You need to explicitly call :meth:`get_secret_value()` to get the wrapped
-    value.  Thus, it is no drop-in replacement for the wrapped data.
+    You need to explicitly call :meth:`get_secret_value()` to get the wrapped value.
+    Thus, it is no drop-in replacement for the wrapped data.
 
-    See :class:`SecretStr` if you need a drop-in replacement for strings,
-    even if it is not quite as safe.
+    See :class:`SecretStr` if you need a drop-in replacement for strings, even if it is
+    not quite as safe.
 
-    You can use :class:`bool` to get the boolean value of the wrapped secret.
-    Other protocol methods (e.g., for length or comparison operators) are not
-    implemented.
+    You can use :class:`bool` to get the boolean value of the wrapped secret.  Other
+    protocol methods (e.g., for length or comparison operators) are not implemented.
 
     .. versionadded:: 2.0.0
     """